added a base class and inherited others from it
