+++ src/test/java/autotests/DuckCreateTest.java
@@ -12,7 +12,7 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 
-public class DuckCreateTest extends TestNGCitrusSpringSupport {
+public class DuckCreateTest extends DuckBaseTest {
     @Test(description = "Проверка создания утки")
     @CitrusTest
     public void successfulCreateDucks(@Optional @CitrusResource TestCaseRunner runner){
@@ -24,41 +24,10 @@ public class DuckCreateTest extends TestNGCitrusSpringSupport {
     }
 
     private void createDuckAndValidate(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        // Создаем утку
+
         createDuck(runner, color, height, material, sound, wingsState);
-
-        // Сохраняем ID утки
         saveDuckId(runner);
-
-        // Проверяем ответ
         validateDuckResponse(runner, color, height, material, sound, wingsState);
-    }
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-    private void saveDuckId(TestCaseRunner runner) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id", "duckId")) // Сохраняем ID в переменной duckId
-        );
     }
 
     private void validateDuckResponse(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
