+++ src/test/java/autotests/DuckDeleteTest.java
@@ -12,7 +12,7 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 
-public class DuckDeleteTest extends TestNGCitrusSpringSupport {
+public class DuckDeleteTest extends DuckBaseTest {
     @Test(description = "Проверка удаления утки")
     @CitrusTest
     public void successfulDeleteDuck(@Optional @CitrusResource TestCaseRunner runner){
@@ -25,39 +25,6 @@ public class DuckDeleteTest extends TestNGCitrusSpringSupport {
     private void createDuckAndSaveId(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
         createDuck(runner, color, height, material, sound, wingsState);
         saveDuckId(runner);
-    }
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-    private void saveDuckId(TestCaseRunner runner) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id", "duckId")) // Сохраняем ID в переменной duckId
-        );
-    }
-
-    private void deleteDuck(TestCaseRunner runner, String duckId) {
-        runner.$(
-                http().client("http://localhost:2222").send().delete("/api/duck/delete").queryParam("id", duckId)
-        );
     }
 
     private void checkDuckDeleted(TestCaseRunner runner, String duckId) {
