+++ src/test/java/autotests/DuckFlyTest.java
@@ -12,7 +12,7 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 
-public class DuckFlyTest extends TestNGCitrusSpringSupport {
+public class DuckFlyTest extends DuckBaseTest {
 
     @Test (description = "Проверка того, что утка летает")
     @CitrusTest
@@ -33,33 +33,6 @@ public class DuckFlyTest extends TestNGCitrusSpringSupport {
         validateFlyResponse(runner, expectedMessage);
     }
 
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-    private void saveDuckId(TestCaseRunner runner){
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id","duckId"))
-        );
-    }
 
     public void duckFly(TestCaseRunner runner, String id){
         runner.$(
