+++ src/test/java/autotests/DuckPropertiesTest.java
@@ -12,7 +12,7 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 
-public class DuckPropertiesTest extends TestNGCitrusSpringSupport {
+public class DuckPropertiesTest extends DuckBaseTest {
     @Test(description = "Проверка того, что выводятся свойства уточки")
     @CitrusTest
     public void successfulProperties(@Optional @CitrusResource TestCaseRunner runner){
@@ -31,34 +31,6 @@ public class DuckPropertiesTest extends TestNGCitrusSpringSupport {
     public void getDuckProperties(TestCaseRunner runner, String id){
         runner.$(
                 http().client("http://localhost:2222").send().get("/api/duck/action/properties").queryParam("id", id)
-        );
-    }
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-
-    private void saveDuckId(TestCaseRunner runner){
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id","duckId"))
         );
     }
