+++ src/test/java/autotests/DuckQuackTest.java
@@ -12,7 +12,7 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 
-public class DuckQuackTest extends TestNGCitrusSpringSupport {
+public class DuckQuackTest extends DuckBaseTest {
     @Test (description = "Проверка того, что утки с четным и нечетным id издают звуки")
     @CitrusTest
     public void successfulQuack(@Optional @CitrusResource TestCaseRunner runner){
@@ -30,33 +30,6 @@ public class DuckQuackTest extends TestNGCitrusSpringSupport {
     private void createDuckAndSaveId(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
         createDuck(runner, color, height, material, sound, wingsState);
         saveDuckId(runner);
-    }
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-    private void saveDuckId(TestCaseRunner runner) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id", "duckId")) // Сохраняем ID в переменной duckId
-        );
     }
 
     private void checkDuckQuack(TestCaseRunner runner, String duckId, int repetitionCount, int soundCount, String expectedSound) {
