+++ src/test/java/autotests/DuckSwimTest.java
@@ -13,7 +13,7 @@ import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.validation.DelegatingPayloadVariableExtractor.Builder.fromBody;
 import static com.consol.citrus.variable.MessageHeaderVariableExtractor.Builder.fromHeaders;
 
-public class DuckSwimTest extends TestNGCitrusSpringSupport {
+public class DuckSwimTest extends DuckBaseTest {
     @Test(description = "Проверка того, что уточка поплыла")
     @CitrusTest
     public void successfulSwim(@Optional @CitrusResource TestCaseRunner runner){
@@ -25,17 +25,6 @@ public class DuckSwimTest extends TestNGCitrusSpringSupport {
         duckSwim(runner, "-1");
         validateBadResponse(runner);
         extractDataFromResponse(runner);
-    }
-
-    private void saveDuckId(TestCaseRunner runner){
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response()
-                        .message()
-                        .extract(fromBody().expression("$.id","duckId"))
-        );
     }
 
     public void duckSwim(TestCaseRunner runner, String id){
@@ -56,22 +45,6 @@ public class DuckSwimTest extends TestNGCitrusSpringSupport {
                 http().client("http://localhost:2222").receive().response(HttpStatus.NOT_FOUND).
                         message().contentType(MediaType.APPLICATION_JSON_VALUE)
         );
-    }
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body( "{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
     }
 
     public void extractDataFromResponse(TestCaseRunner runner) {
